Introduction
Slightly revised text in the introduction. I have taken out most of
Tihana’s paragraph because I don’t understand it very well; if we want
it then we need to clarify some of the statements about other
techniques for specifying communication. All the old text is in
intro-OLD.tex